+++ src/GitHub_Updater/Bitbucket_API.php
@@ -35,6 +35,7 @@ class Bitbucket_API extends API implements API_Interface {
 	 * @param \stdClass $type The repo type.
 	 */
 	public function __construct( $type ) {
+		parent::__construct();
 		$this->type     = $type;
 		$this->response = $this->get_repo_cache();
 		$branch         = new Branch( $this->response );