+++ src/GitHub_Updater/Install.php
@@ -41,8 +41,8 @@ class Install extends Base {
 	 * @param array  $wp_cli_config
 	 */
 	public function __construct( $type, $wp_cli_config = array() ) {
+		parent::__construct();
 		require_once ABSPATH . 'wp-admin/includes/class-wp-upgrader.php';
-		$this->install( $type, $wp_cli_config );
 
 		wp_enqueue_script( 'ghu-install', plugins_url( basename( dirname( dirname( __DIR__ ) ) ) . '/js/ghu_install.js' ), array(), false, true );
 	}
@@ -55,12 +55,12 @@ class Install extends Base {
 	 *
 	 * @return bool
 	 */
-	public function install( $type, $wp_cli_config ) {
+	public function install( $type, $wp_cli_config = null ) {
 		$wp_cli = false;
 
 		if ( ! empty( $wp_cli_config['uri'] ) ) {
 			$wp_cli  = true;
-			$headers = Base::parse_header_uri( $wp_cli_config['uri'] );
+			$headers = $this->parse_header_uri( $wp_cli_config['uri'] );
 			$api     = false !== strpos( $headers['host'], '.com' )
 				? rtrim( $headers['host'], '.com' )
 				: rtrim( $headers['host'], '.org' );
@@ -106,7 +106,7 @@ class Install extends Base {
 			/*
 			 * Transform URI to owner/repo
 			 */
-			$headers                      = Base::parse_header_uri( $_POST['github_updater_repo'] );
+			$headers                      = $this->parse_header_uri( $_POST['github_updater_repo'] );
 			$_POST['github_updater_repo'] = $headers['owner_repo'];
 
 			self::$install         = Settings::sanitize( $_POST );