+++ src/GitHub_Updater/Theme.php
@@ -42,13 +42,14 @@ class Theme extends Base {
 	 * Constructor.
 	 */
 	public function __construct() {
+		parent::__construct();
 
 		/*
 		 * Get details of installed git sourced themes.
 		 */
 		$this->config = $this->get_theme_meta();
 
-		if ( empty( $this->config ) ) {
+		if ( null === $this->config ) {
 			return;
 		}
 	}