+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+* removed non-constructor stuff from constructors
+* added `parent::__construct()` to all extended classes
 
 #### 7.2.0 / 2017-08-30
 * added a static proxy class to use for creating Singletons